chore(SPV-750): move PAgeResponse struct to models
