+++ actions/common/search_methods.go
@@ -4,6 +4,7 @@ import (
 	"math"
 
 	"github.com/bitcoin-sv/spv-wallet/engine/datastore"
+	"github.com/bitcoin-sv/spv-wallet/models"
 )
 
 // LoadDefaultQueryParams will load the default query parameters
@@ -15,9 +16,9 @@ func LoadDefaultQueryParams() *datastore.QueryParams {
 }
 
 // GetPageFromQueryParams will return a Page object from the query parameters and count value
-func GetPageFromQueryParams(queryParams *datastore.QueryParams, count int64) Page {
+func GetPageFromQueryParams(queryParams *datastore.QueryParams, count int64) models.Page {
 	totalPages := int(math.Ceil(float64(count) / float64(queryParams.PageSize)))
-	page := Page{
+	page := models.Page{
 		Size:          queryParams.PageSize,
 		Number:        queryParams.Page,
 		TotalElements: count,
