+++ actions/contacts/search.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/bitcoin-sv/spv-wallet/actions/common"
 	"github.com/bitcoin-sv/spv-wallet/mappings"
+	"github.com/bitcoin-sv/spv-wallet/models"
 	"github.com/bitcoin-sv/spv-wallet/server/auth"
 	"github.com/gin-gonic/gin"
 )
@@ -65,7 +66,7 @@ func (a *Action) search(c *gin.Context) {
 		return
 	}
 
-	response := common.SearchContactsResponse{
+	response := models.SearchContactsResponse{
 		Content: contracts,
 		Page:    common.GetPageFromQueryParams(reqParams.QueryParams, count),
 	}
